test(utils): fold the Unicode 17 case tests into case_test.go
They test StringToUppercase and StringToLowercase, so they belong beside the
rest of case.go's tests. The separate file was named after unicode17.go back
when that sat in this package; now that the delta is a package of its own,
the name only collides with the test file that package carries.
